- Fix some more compilation warnings detected by VC10 on Windows 64 bits

+++ api/cpp/cppapi/branches/NextRelease/client/dbapi.h
@@ -556,7 +556,7 @@ public :
 	DbDatum(const DbDatum &);
 	DbDatum &operator=(const DbDatum &);
 
-	int size() {return value_string.size();}
+	size_t size() {return value_string.size();}
 	bool is_empty();
 
 	void exceptions(bitset<numFlags> fl) { exceptions_flags = fl;}
